+++ test/model/relationship/has_many_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_data/flutter_data.dart';
 import 'package:test/test.dart';
 
-import '../../../models/family.dart';
-import '../../../models/person.dart';
-import '../../setup.dart';
+import '../../_support/family.dart';
+import '../../_support/person.dart';
+import '../../_support/setup.dart';
 
 void main() async {
   setUp(setUpFn);
