+++ test/model/relationship/relationship_test.dart
@@ -1,12 +1,12 @@
 import 'package:flutter_data/flutter_data.dart';
 import 'package:test/test.dart';
 
-import '../../../models/family.dart';
-import '../../../models/house.dart';
-import '../../../models/node.dart';
-import '../../../models/person.dart';
-import '../../../models/pet.dart';
-import '../../setup.dart';
+import '../../_support/family.dart';
+import '../../_support/house.dart';
+import '../../_support/node.dart';
+import '../../_support/person.dart';
+import '../../_support/pet.dart';
+import '../../_support/setup.dart';
 
 void main() async {
   setUp(setUpFn);
